# docs/reference/ui-api.openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Hello": {
        "additionalProperties": false,
        "properties": {
          "api_instance": {
            "type": "string"
          },
          "api_version": {
            "type": "integer"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "server_version": {
            "type": "string"
          },
          "terminal": {
            "$ref": "#/components/schemas/TerminalProtocol"
          }
        },
        "required": [
          "api_version",
          "api_instance",
          "server_version",
          "capabilities",
          "terminal"
        ],
        "type": "object"
      },
...
      "OriginList": {
        "additionalProperties": false,
        "properties": {
          "origins": {
            "items": {
              "$ref": "#/components/schemas/PairedOrigin"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "origins"
        ],
        "type": "object"
      },
...
      "OriginRevocation": {
        "additionalProperties": false,
        "properties": {
          "origin": {
            "type": "string"
          },
          "revoked": {
            "type": "boolean"
          }
        },
        "required": [
          "origin",
          "revoked"
        ],
        "type": "object"
      },
...
      "PairingCodeRequest": {
        "additionalProperties": false,
        "properties": {
          "next": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "tailnetLogin": {
        "description": "Trusted only on the dedicated Tailnet listener.",
        "in": "header",
        "name": "Tailscale-User-Login",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Generated by sidecar api spec. HTTP and WebSocket auth, guards and ordering: docs/reference/ui-api.md and mobile-protocol.md.",
    "title": "Sidecar UI API",
    "version": "0"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v0/hello": {
      "get": {
        "operationId": "hello_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hello"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "tailnetLogin": []
          }
        ],
        "x-listeners": [
          "local",
          "browser",
          "tailnet"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      },
      "head": {
        "operationId": "hello_get_head",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "tailnetLogin": []
          }
        ],
        "x-listeners": [
          "local",
          "browser",
          "tailnet"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      }
    },
    "/api/v0/origins": {
      "delete": {
        "operationId": "origins_delete",
        "parameters": [
          {
            "description": "Required on Browser and Tailnet; not on Local.",
            "in": "header",
            "name": "X-Sidecar-Request",
            "required": false,
            "schema": {
              "const": "1",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OriginRevocation"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {}
        ],
        "x-listeners": [
          "local"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      },
      "get": {
        "operationId": "origins_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OriginList"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {}
        ],
        "x-listeners": [
          "local"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      },
      "head": {
        "operationId": "origins_get_head",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {}
        ],
        "x-listeners": [
          "local"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      },
      "post": {
        "operationId": "origins_post",
        "parameters": [
          {
            "description": "Required on Browser and Tailnet; not on Local.",
            "in": "header",
            "name": "X-Sidecar-Request",
            "required": false,
            "schema": {
              "const": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OriginRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OriginRegistration"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {}
        ],
        "x-listeners": [
          "local"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      }
    },
    "/api/v0/pairing/codes": {
      "post": {
        "operationId": "pairing_codes_post",
        "parameters": [
          {
            "description": "Required on Browser and Tailnet; not on Local.",
            "in": "header",
            "name": "X-Sidecar-Request",
            "required": false,
            "schema": {
              "const": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingCodeRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCode"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {}
        ],
        "x-listeners": [
          "local"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      }
    },
    "/api/v0/pairing/exchange": {
      "post": {
        "operationId": "pairing_exchange_post",
        "parameters": [
          {
            "description": "Required on Browser and Tailnet; not on Local.",
            "in": "header",
            "name": "X-Sidecar-Request",
            "required": false,
            "schema": {
              "const": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingExchangeRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingExchange"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [],
        "x-listeners": [
          "browser"
        ]
      }
    },
    "/api/v0/pairing/sessions": {
      "delete": {
        "operationId": "pairing_sessions_delete",
        "parameters": [
          {
            "description": "Required on Browser and Tailnet; not on Local.",
            "in": "header",
            "name": "X-Sidecar-Request",
            "required": false,
            "schema": {
              "const": "1",
              "type": "string"
            }
          },
          {
            "description": "Omit to revoke every browser session; supply an origin to revoke only its sessions.",
            "in": "query",
            "name": "origin",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRevocation"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Named refusal; see ui-api.md for status codes and remedies"
          }
        },
        "security": [
          {}
        ],
        "x-listeners": [
          "local"
        ],
        "x-local-auth": "No credential on the mode-0600 Unix socket; the empty security alternative applies only there."
      }
    },
...
  },
...
}